backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
pymongo==4.5.0
mongoengine==0.27.0
python-dotenv==1.0.0
bcrypt==4.0.1
Werkzeug==2.3.7
marshmallow==3.20.1
flask-marshmallow==0.15.0
Pillow==10.0.1
python-dateutil==2.8.2
email-validator==2.0.0
gunicorn==21.2.0